% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboveavr.R
\name{aboveavr}
\alias{aboveavr}
\title{Movies with above average budgets}
\usage{
aboveavr(n)
}
\arguments{
\item{n}{Year / must be in between [2009,2011]}
}
\value{
Movie names
}
\description{
This function lists the movies which have above average budgets for a selected year
}
\examples{

aboveavr ()
aboveavr (2009)

}
