% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_ratings.R
\name{boxplot_ratings}
\alias{boxplot_ratings}
\title{Critics Ratings vs Audience Ratings Box Plots}
\usage{
boxplot_ratings()
}
\value{
Box Plots
}
\description{
Displays Critics Ratings vs. Audience Ratings by genre
}
