% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corre.R
\name{corre}
\alias{corre}
\title{Correlation between Critics Rating and Budget}
\usage{
corre()
}
\value{
correlation number between -1 and 1, and comment
}
\description{
Compute the correlation between Critics Ratings and Budget and tell how strong the computed correlation is.
}
