% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critvsaud.R
\name{critvsaud}
\alias{critvsaud}
\title{Critics Ratings vs Audience Ratings Line Plot}
\usage{
critvsaud()
}
\value{
Point Line Plot
}
\description{
Displays Critics Ratings vs. Audience Ratings by genre and budget
}
