% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fltmovie.R
\name{fltmovie}
\alias{fltmovie}
\title{Filtering top rated movies by genre, budget and year}
\usage{
fltmovie(gnr, n, yr)
}
\arguments{
\item{gnr}{genre / "in quotations"}

\item{n}{budget amount}

\item{yr}{year / between 2007-2011}
}
\value{
movie names
}
\description{
This functions gives you the movies which have higher critics ratings than
\%90 of the data
}
\examples{
\dontrun{
fltmovie("genre", budget in millions, year)
fltmovie ("Action",100,2009)

}

}
