% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovieDataAnalysis.R
\name{MovieDataAnalysis}
\alias{MovieDataAnalysis}
\title{Movie Ratings Analysis}
\description{
Contains functions for both analyzing and visualizing movie ratings data
}
\author{
Nazli Begum Cirpanli \ e-mail {nazlibegum.cirpanli@studenti.unimi.it}
}
