% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingsmode.R
\name{ratingsmode}
\alias{ratingsmode}
\title{Ratings Mode}
\usage{
ratingsmode()
}
\value{
A vector with column names
}
\description{
This function computes the most recurring ratings for Critics and Audience Ratings
over the five years
}
\examples{
\dontrun{
ratingsmode()

}

}
