% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyavrgenrebdg.R
\name{yearlyavrgenrebdg}
\alias{yearlyavrgenrebdg}
\title{Average budget by genre for selected year}
\usage{
yearlyavrgenrebdg(n)
}
\arguments{
\item{n}{Year / must be in between \link{2009,2011}}
}
\value{
A data frame with genres and average budget amounts
}
\description{
This function computes the average budget amounts by genres for a given year
}
\examples{
 \dontrun{
yearlyavrgenrebdg()
yearlyavrgenrebdg(2010)

}
}
